Refactor BIOSTEST out of dos_programs
