+++ src/dos/dos_programs.cpp
@@ -59,69 +59,10 @@
 #include "../ints/int10.h"
 
 #if C_DEBUG
-Bitu DEBUG_EnableDebugger(void);
+	#include "program_biostest.h"
 #endif
 
 extern Bit32u floppytype;
-
-#if C_DEBUG
-class BIOSTEST final : public Program {
-public:
-	void Run(void) {
-		if (!(cmd->FindCommand(1, temp_line))) {
-			WriteOut("Must specify BIOS file to load.\n");
-			return;
-		}
-
-		Bit8u drive;
-		char fullname[DOS_PATHLENGTH];
-		localDrive* ldp = 0;
-		if (!DOS_MakeName((char *)temp_line.c_str(), fullname, &drive)) return;
-
-		try {
-			/* try to read ROM file into buffer */
-			ldp = dynamic_cast<localDrive*>(Drives[drive]);
-			if (!ldp) return;
-
-			FILE *tmpfile = ldp->GetSystemFilePtr(fullname, "rb");
-			if (tmpfile == NULL) {
-				WriteOut("Can't open a file");
-				return;
-			}
-			fseek(tmpfile, 0L, SEEK_END);
-			if (ftell(tmpfile) > 64 * 1024) {
-				WriteOut("BIOS File too large");
-				fclose(tmpfile);
-				return;
-			}
-			fseek(tmpfile, 0L, SEEK_SET);
-			Bit8u buffer[64*1024];
-			Bitu data_read = fread(buffer, 1, sizeof( buffer), tmpfile);
-			fclose(tmpfile);
-
-			Bit32u rom_base = PhysMake(0xf000, 0); // override regular dosbox bios
-			/* write buffer into ROM */
-			for (Bitu i = 0; i < data_read; i++) phys_writeb(rom_base + i, buffer[i]);
-
-			//Start executing this bios
-			memset(&cpu_regs, 0, sizeof(cpu_regs));
-			memset(&Segs, 0, sizeof(Segs));
-
-
-			SegSet16(cs, 0xf000);
-			reg_eip = 0xfff0;
-		}
-		catch (...) {
-			return;
-		}
-	}
-};
-
-static void BIOSTEST_ProgramStart(Program * * make) {
-	*make = new BIOSTEST;
-}
-
-#endif
 
 void DOS_SetupPrograms(void) {
 	/*Add Messages */
